# Conflicts:
#	code/game/objects/structures/light_pole.dm
#	code/game/objects/structures/urban.dm
#	icons/obj/signs.dmi
#	icons/obj/structure/urban/restaurant.dmi
#	maps/away/away_site/konyang/point_verdant/point_verdant-1.dmm
#	maps/away/away_site/konyang/point_verdant/point_verdant-2.dmm
#	maps/away/away_site/konyang/point_verdant/point_verdant-3.dmm
#	maps/event/halloween/halloween.dmm
#	maps/event/idris_cruise/idris_cruise-1.dmm
#	maps/event/rooftop/rooftop-1.dmm
#	maps/random_ruins/exoplanets/konyang/abandoned/abandoned_outpost.dmm
#	maps/random_ruins/exoplanets/konyang/abandoned/house_small.dmm
#	maps/random_ruins/exoplanets/konyang/abandoned/landing_zone.dmm
#	maps/random_ruins/exoplanets/konyang/homestead.dmm
#	maps/random_ruins/exoplanets/konyang/landing_zone.dmm
#	maps/random_ruins/exoplanets/konyang/ocean/naval_landing_zone.dmm
#	maps/random_ruins/exoplanets/konyang/telecomms_outpost.dmm
#	maps/random_ruins/exoplanets/konyang/village.dmm
#	maps/sccv_horizon/sccv_horizon-4_centcomm.dmm

+++ code/game/objects/structures/curtains.dm
@@ -1,11 +1,8 @@
-#define SHOWER_OPEN_LAYER OBJ_LAYER + 0.4
-#define SHOWER_CLOSED_LAYER MOB_LAYER + 0.1
-
 /obj/structure/curtain
 	name = "curtain"
 	icon = 'icons/obj/curtain.dmi'
 	icon_state = "closed"
-	layer = SHOWER_OPEN_LAYER
+	layer = ABOVE_WINDOW_LAYER
 	opacity = 1
 	density = 0
 	anchored = TRUE //curtains start secured in place
@@ -20,15 +17,17 @@
 
 /obj/structure/curtain/open
 	icon_state = "open"
-	layer = SHOWER_CLOSED_LAYER
+	layer = ABOVE_HUMAN_LAYER
 	opacity = 0
 
-/obj/structure/curtain/bullet_act(obj/item/projectile/P, def_zone)
-	if(!P.nodamage)
-		visible_message("<span class='warning'>[P] tears [src] down!</span>")
+/obj/structure/curtain/bullet_act(obj/projectile/hitting_projectile, def_zone, piercing_hit)
+	. = ..()
+	if(. != BULLET_ACT_HIT)
+		return .
+
+	if(hitting_projectile.damage > 0)
+		visible_message(SPAN_WARNING("[hitting_projectile] tears [src] down!"))
 		qdel(src)
-	else
-		..(P, def_zone)
 
 /obj/structure/curtain/attack_hand(mob/user)
 	playsound(get_turf(loc), 'sound/effects/curtain.ogg', 15, 1, -5)
@@ -60,17 +59,17 @@
 			if(C != src && C.anchored) //Can't secure more than one curtain in a tile
 				to_chat(user, "There is already a curtain secured here!")
 				return
-		playsound(src.loc, attacking_item.usesound, 50, 1)
+		attacking_item.play_tool_sound(get_turf(src), 50)
 		visible_message(SPAN_NOTICE("\The [src] has been [anchored ? "secured in place" : "unsecured"] by \the [user]."))
 
 /obj/structure/curtain/proc/toggle()
 	src.set_opacity(!src.opacity)
 	if(opacity)
 		icon_state = "closed"
-		layer = SHOWER_CLOSED_LAYER
+		layer = ABOVE_HUMAN_LAYER
 	else
 		icon_state = "open"
-		layer = SHOWER_OPEN_LAYER
+		layer = ABOVE_WINDOW_LAYER
 
 /obj/structure/curtain/black
 	name = "black curtain"
@@ -109,6 +108,3 @@
 
 /obj/structure/curtain/open/shower/security
 	color = "#AA0000"
-
-#undef SHOWER_OPEN_LAYER
-#undef SHOWER_CLOSED_LAYER
